% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{canopyexplorer}
\alias{canopyexplorer}
\title{Canopy Explorer}
\description{
The \code{canopyexplorer} package includes functions for cleaning, analyzing,
and visualizing Canopy school survey data. The package provides tools
for loading publicly available datasets, using Canopy branding, running
analysis, and more.
}
\keyword{package}
