% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{clean_locale}
\alias{clean_locale}
\title{Cleaning function for Canopy school data}
\usage{
clean_locale(data, urban, rural, suburban)
}
\arguments{
\item{data}{A data frame.}

\item{urban}{A string indicating the name of the column that specifies if a
school serves students in urban locales.}

\item{rural}{A string indicating the name of the column that specifies if a
school serves students in rural locales.}

\item{suburban}{A string indicating the name of the column that specifies if a
school serves students in suburban locales.}
}
\value{
A data frame with a new \code{locale} column summarizing the locale type.
Possible values are \code{"Urban"}, \code{"Suburban"}, \code{"Rural"}, \code{"Multiple"}, or \code{NA}.
}
\description{
This function takes three numeric binary columns (0/1) indicating whether a
school serves urban, suburban, and rural locales, and collapses them into
a single categorical column labeled \code{locale}.
}
\details{
The function assumes that the input binary columns are coded as \code{0} or \code{1}.
If multiple binary indicators are \code{1}, the function will classify the school
as \code{"Multiple"}. If none of the indicators are \code{1}, the value will be \code{NA}.
\itemize{
\item If any binary column is not coded as \code{0/1}, a warning will be issued and
the value will be coerced to \code{NA}.
\item If multiple indicators are set to \code{1}, the function will return \code{"Multiple"}.
\item If all indicators are \code{0} or \code{NA}, the function will return \code{NA}.
}
}
\note{
If a \code{locale} column already exists in the input data, it will be
overwritten.
}
\examples{
# Example data
test_data <- data.frame(
  urban = c(1, 0, 0, 1, 0),
  rural = c(0, 1, 0, 1, 0),
  suburban = c(0, 0, 1, 0, 1)
)

clean_locale(test_data, "urban", "rural", "suburban")

}
\seealso{
\code{\link{load_canopy_data}} for loading Canopy datasets,
\code{\link{clean_practices}} for transforming columns documenting practices,
\code{\link{clean_grades}} for creating a column summarizing school level
}
