% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{clean_practices}
\alias{clean_practices}
\title{Cleaning function for tag data}
\usage{
clean_practices(data, prefix = "practices_")
}
\arguments{
\item{data}{A data frame containing the tag data.}

\item{prefix}{A character string indicating the prefix for the tag variables,
default \code{"practices_"}.}
}
\value{
A data frame with the new \verb{practices_*}, \verb{core_*}, and \verb{time_*} columns
added.
}
\description{
A cleaning function that takes the original values for all tags with a
specified prefix and returns 3 cleaned versions of the tag:
\itemize{
\item \code{"practices_*"} — A binary variable indicating if a school selected a tag.
\item \code{"core_*"} — A binary variable indicating a school selected a tag and
indicated the practice is core to their school's design.
\item \code{"time_*"} — A variable indicating the length of time the core practice
has been in use at a schoolwide level.
}
}
\details{
\itemize{
\item This function creates three new columns for each original tag:
\itemize{
\item \code{"practices_*"} is \code{1} if the practice was selected, \code{0} otherwise.
\item \code{"core_*"} is \code{1} if the practice was marked as core to the school's design.
\item \code{"time_*"} reflects the length of time the core practice has been in place.
}
\item If the original tag column contains \code{NA}, the output will be \code{NA}.
}
}
\note{
If output columns already exist, they will be overwritten.
}
\examples{
test_data <- data.frame(
  practices_math = c("1-2 years", "Less than a year", "Not sure", NA, "5+ years"),
  practices_reading = c("3-4 years", "5+ years", "Less than a year", "1", "Not sure")
)

clean_practices(test_data, prefix = "practices_")

}
\seealso{
\code{\link{load_canopy_data}} for loading Canopy datasets,
\code{\link{clean_locale}} for creating a column summarizing school geographic locale,
\code{\link{clean_grades}} for creating a column summarizing school level
}
