% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_crpe_functions.R
\name{crpe_palette}
\alias{crpe_palette}
\title{CRPE color palettes for ggplot objects}
\usage{
crpe_palette(palette_name, type = c("fill", "color"), secondary_palette = NULL)
}
\arguments{
\item{palette_name}{A character string indicating the primary color palette to use.
Options include:
\itemize{
\item \code{"crpe_blues"}
\item \code{"crpe_oranges"}
\item \code{"crpe_purples"}
\item \code{"crpe_grays"}
}}

\item{type}{A character string indicating whether to apply the palette to the
fill or color aesthetic. Options are \code{"fill"} or \code{"color"}. Default is \code{"fill"}.}

\item{secondary_palette}{A character string indicating the secondary palette to use
if the primary palette runs out of colors. If \code{NULL}, the primary palette will be
recycled. Options are the same as \code{palette_name}.}
}
\value{
A ggplot2 scale object.
}
\description{
A wrapper function to make CRPE color palettes compatible with ggplot
objects when using categorical data. Supports both fill and color aesthetics.
Allows for a secondary palette if the primary palette does not have enough colors.
}
\examples{
# Use for fill aesthetic with overflow handled by secondary palette
ggplot(mpg, aes(class, fill = drv)) +
  geom_bar() +
  crpe_palette("crpe_blues", type = "fill", secondary_palette = "crpe_oranges")

# Use for color aesthetic
ggplot(mpg, aes(displ, hwy, color = drv)) +
  geom_point() +
  crpe_palette("crpe_oranges", type = "color")

}
