% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_crpe_functions.R
\name{expand_axes}
\alias{expand_axes}
\title{Expand axes for ggplot objects}
\usage{
expand_axes(expand_x = FALSE, expand_y = FALSE)
}
\arguments{
\item{expand_x}{Logical. If \code{TRUE}, removes padding from the x-axis when using
continuous variables. Cannot be used with discrete variables. Default is \code{FALSE}.}

\item{expand_y}{Logical. If \code{TRUE}, removes padding from the y-axis when using
continuous variables. Cannot be used with discrete variables. Default is \code{FALSE}.}
}
\value{
A list of ggplot scale objects that can be added to a ggplot object.
}
\description{
This function removes padding from the x and/or y axes in ggplot objects
when using continuous variables. It is designed to be used in combination
with \code{theme_crpe()} to adjust the axis expansion behavior.
}
\examples{
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  theme_crpe() +
  expand_crpe_axes(expand_x = TRUE, expand_y = TRUE)

}
\seealso{
\code{\link{theme_crpe}} for the CRPE theme.
}
