% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_canopy_data.R
\name{load_canopy_data}
\alias{load_canopy_data}
\title{Load Canopy datasets}
\usage{
load_canopy_data(year)
}
\arguments{
\item{year}{specifies which year's data you would like to load. Options
currently include 2019, 2021a, 2021b, 2022, 2023, or 2024. Note: there is no
data set for 2020 due to COVID. 2021a represents the typical spring school
survey, and 2021b represents a follow-up survey that Canopy conducted to
track rapid changes during COVID-era school responses.}
}
\value{
A data frame containing the selected Canopy dataset.
}
\description{
This is a function to load publicly-available Canopy datasets. These datasets
can also be found and manually downloaded at canopyschools.org.
}
\details{
\itemize{
\item Data structure may vary slightly between years due to changes in survey
design and reporting conventions.
}
}
