% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_functions.R
\name{tag_bayes}
\alias{tag_bayes}
\title{Run Bayesian GLM models for tags}
\usage{
tag_bayes(data, covariates, exponentiate = FALSE)
}
\arguments{
\item{data}{A dataframe of prepped data from \code{tag_bayes_prep}.}

\item{covariates}{A vector of covariates to include in the model.}

\item{exponentiate}{Logical. If \code{TRUE}, exponentiates the output to show odds ratios
rather than log-odds. Default is \code{FALSE}.}
}
\value{
A tidy dataframe of model output.
}
\description{
This function runs a Bayesian GLM model for each tag using prepped Canopy data
and specified covariates. The model is fitted using \code{rstanarm::stan_glm()}
and returns tidy results.
}
\seealso{
\code{\link{tag_bayes_prep}} to prepare data for modeling.
\code{\link{tag_bayes_plot}} to plot the model results.
}
