% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_functions.R
\name{tag_bayes_plot}
\alias{tag_bayes_plot}
\title{Plot results from \code{tag_bayes}}
\usage{
tag_bayes_plot(data, tag = NULL, n = 10, exponentiate = FALSE, theme = "crpe")
}
\arguments{
\item{data}{A data frame of model results created by \code{tag_bayes}.}

\item{tag}{A specific tag to plot. If \code{NULL}, the tag with the largest
absolute effect will be selected automatically. Default is \code{NULL}.}

\item{n}{Number of terms to show in the plot. Default is \code{10}.}

\item{theme}{Character; \code{"crpe"}, \code{"transcend"}, or \code{"minimal"} to apply a
specific theme to the plot. Default is \code{"crpe"}.}
}
\value{
A ggplot object showing the top \code{n} absolute effects for the specified tag.
}
\description{
A function that creates a plot of the top \code{n} absolute effects for a selected
tag from the results of \code{tag_bayes}. If no tag is specified, the function
will select the tag with the largest absolute effect.
}
\seealso{
\code{\link{tag_bayes}} to create the model results used in this plot.
}
