% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_functions.R
\name{tag_bayes_prep}
\alias{tag_bayes_prep}
\title{Prepare data for Bayesian modeling}
\usage{
tag_bayes_prep(data, prefix = "practices_", id = "school_id", covariates)
}
\arguments{
\item{data}{A dataframe of Canopy data including tags and covariates.}

\item{prefix}{The prefix used for the tags to identify relevant variables.}

\item{id}{A column of unique identifiers for schools. Default is \code{"school_id"}.}

\item{covariates}{A vector of covariates to include in the model.}
}
\value{
A prepped dataframe ready for modeling.
}
\description{
This function prepares Canopy data for modeling by pivoting wide tag columns
into long format and scaling numeric covariates specified.
}
\seealso{
\code{\link{tag_bayes}} to run Bayesian models on the prepped data.
}
