% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_crpe_functions.R
\name{theme_crpe}
\alias{theme_crpe}
\title{CRPE style theme for ggplot objects}
\usage{
theme_crpe(percent_labels = FALSE)
}
\arguments{
\item{percent_labels}{Logical. Whether the y-axis should display percentage
labels (rounded to the nearest whole number). Default is \code{FALSE}.}

\item{expand_x}{Logical. Whether x-axis padding should be removed when
using continuous variables. Cannot be used with discrete variables.
Default is \code{FALSE}.}

\item{expand_y}{Logical. Whether y-axis padding should be removed when
using continuous variables. Cannot be used with discrete variables.
Default is \code{FALSE}.}

\item{x_var}{A vector representing the x-axis data. Only used to validate
whether \code{expand_x} is applicable.}

\item{y_var}{A vector representing the y-axis data. Only used to validate
whether \code{expand_y} is applicable.}
}
\value{
A ggplot2 theme object.
}
\description{
This is a ggplot2 theme that applies CRPE branding to a plot, including
custom fonts, expanded margins, and removal of minor axis lines and padding.
It is designed for bar plots, scatter plots, and line plots.
}
\examples{
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  theme_crpe(expand_x = TRUE, expand_y = TRUE)

}
