% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_crpe_functions.R
\name{theme_crpe_sparse}
\alias{theme_crpe_sparse}
\title{Minimal CRPE style theme for ggplot objects}
\usage{
theme_crpe_sparse()
}
\value{
A ggplot2 theme object.
}
\description{
This is a simplified ggplot2 theme that applies CRPE branding while
removing all axis lines and simplifying the plot's visual appearance.
It is intended for minimalist plots where gridlines and borders
are unnecessary.
}
\examples{
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  theme_crpe_sparse()

}
