% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_transcend_functions.R
\name{theme_transcend}
\alias{theme_transcend}
\title{Transcend style theme for ggplot objects}
\usage{
theme_transcend()
}
\value{
A ggplot2 theme object.
}
\description{
This is a ggplot2 theme that applies Transcend branding style,
including expanded margins, specified font, and removed minor axis lines.
It should be used for bar plots, scatter plots, and line plots.
}
\examples{
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  theme_transcend()

}
