% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_transcend_functions.R
\name{theme_transcend_sparse}
\alias{theme_transcend_sparse}
\title{Transcend sparse style theme for ggplot objects}
\usage{
theme_transcend_sparse()
}
\value{
A ggplot2 theme object.
}
\description{
This is a simplified ggplot2 theme that applies Transcend branding style while
removing all axis lines. It is intended for minimalist plots.
}
\examples{
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  theme_transcend_sparse()

}
