% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_transcend_functions.R
\name{transcend_palette}
\alias{transcend_palette}
\title{Transcend color palettes for ggplot objects}
\usage{
transcend_palette(
  palette_name,
  type = c("fill", "color"),
  secondary_palette = NULL
)
}
\arguments{
\item{palette_name}{A character string indicating the color palette to use.
Options include:
\itemize{
\item \code{"transcend_colors"}
\item \code{"transcend_expanded"}
\item \code{"transcend_grays"}
\item \code{"transcend_na"}
}}

\item{type}{A character string indicating whether to apply the palette to the
fill or color aesthetic. Options are \code{"fill"} or \code{"color"}. Default is \code{"fill"}.}

\item{secondary_palette}{Optional. A character string for a secondary palette
if the primary palette runs out of colors.}
}
\value{
A ggplot2 scale object.
}
\description{
A wrapper function to make Transcend color palettes compatible with ggplot
objects when using categorical data. Supports both fill and color aesthetics.
}
\examples{
ggplot(mpg, aes(class, fill = drv)) +
  geom_bar() +
  transcend_palette("transcend_colors", type = "fill", secondary_palette = "transcend_expanded")

}
